requirements/deployment.txt
Fabric==1.10.2
PyYAML==3.11
Jinja2==2.8
